Clamp float-to-int16 audio conversions
This patch makes float-to-int16 conversions on audio samples obtained
from FFmpeg clamp instead of wrap when float samples inexplicably
exceed [-1.0, 1.0], which fixes popping/crackling heard with some
Vorbis-encoded audio, such as the M2 and Infinity menu music
(Music.ogg). (A few pops remain in the M2 music, but these are actually
part of the encoded audio.) The wrapping behavior that causes the pops
was caught by -fsanitize=float-cast-overflow.