autoconf: Use pkg-config for sdl2 (#520)
The correct way to search for sdl2 is pkg-config,
as per https://github.com/libsdl-org/SDL/issues/6140

Replace our legacy `AM_PATH_SDL2` macro with
`AX_REQUIRE_PKG` and update the minimum version
to 2.0.16 to reflect reality.

Drop `m4/` since this was our last macro.

SDL2 is not optional, remove the only `ifdef SDL` block
as without the macro setting that for us we fail
to compile.

Closes: https://github.com/Aleph-One-Marathon/alephone/issues/511

Signed-off-by: Matt Jolly <kangie@gentoo.org>